+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - Drop support for Python 3.7 when using Plone 6.0 @ericof
 
+- Add support to plone.restapi and Volto @ericof
+
 
 ## 5.0.0a2 (2022-09-23)